moves map struct to a mod
